refactor: move queryOwnedPersonaInformation
